+++ src/LibreSpot.Desktop/Converters/LogLevelToBrushConverter.cs
@@ -2,6 +2,9 @@ using System.Globalization;
 using System.Windows;
 using System.Windows.Data;
 using System.Windows.Media;
+using Application = System.Windows.Application;
+using Brush = System.Windows.Media.Brush;
+using Color = System.Windows.Media.Color;
 
 namespace LibreSpot.Desktop.Converters;